fix SPI dependency problem for ESP32-S3
